+++ lapdog/cloud/register.py
@@ -31,16 +31,17 @@ def register(request):
                 400
             )
 
-        if 'token' not in data:
+        token = utils.extract_token(request.headers, data)
+        if token is None:
             return (
                 {
                     'error': 'Bad Request',
-                    'message': 'Missing required parameter "token"'
+                    'message': 'Token must be provided in header or body'
                 },
                 400
             )
 
-        token_data = utils.get_token_info(data['token'])
+        token_data = utils.get_token_info(token)
         if 'error' in token_data:
             return (
                 {
@@ -61,7 +62,7 @@ def register(request):
                 400
             )
 
-        session = utils.generate_user_session(data['token'])
+        session = utils.generate_user_session(token)
 
         read, write = utils.validate_permissions(session, data['bucket'])
         if read is None: