+++ lapdog/cloud/submit.py
@@ -10,7 +10,6 @@ except ImportError:
 import math
 import json
 import traceback
-from dalmatian import getblob
 
 @utils.cors('POST')
 def create_submission(request):
@@ -28,16 +27,17 @@ def create_submission(request):
                 400
             )
 
-        if 'token' not in data:
+        token = utils.extract_token(request.headers, data)
+        if token is None:
             return (
                 {
                     'error': 'Bad Request',
-                    'message': 'Missing required parameter "token"'
+                    'message': 'Token must be provided in header or body'
                 },
                 400
             )
 
-        token_data = utils.get_token_info(data['token'])
+        token_data = utils.get_token_info(token)
         if 'error' in token_data:
             return (
                 {
@@ -72,7 +72,7 @@ def create_submission(request):
                 400
             )
 
-        session = utils.generate_user_session(data['token'])
+        session = utils.generate_user_session(token)
 
         read, write = utils.validate_permissions(session, data['bucket'])
         if read is None:
@@ -264,7 +264,7 @@ def create_submission(request):
 
                 utils.sign_object(
                     (data['submission_id'] + operation).encode(),
-                    getblob(
+                    utils.getblob(
                         'gs://{bucket}/lapdog-executions/{submission_id}/signature'.format(
                             bucket=data['bucket'],
                             submission_id=data['submission_id']